fix: remove test depending on unreleased _pelt_python, fix mypy error
- Replace test_pelt_rust_python_equivalence (imports _pelt_python which
  only exists in the PELT Rust PR) with test_pelt_multiple_shifts
- Cast global_mean to float() in features/advanced.py to fix mypy

+++ polars_ts/features/advanced.py
@@ -39,7 +39,7 @@ def target_encode(
         DataFrame with ``{cat_col}_encoded`` column appended.
 
     """
-    global_mean = df[target_col].mean()
+    global_mean = float(df[target_col].mean())  # type: ignore[arg-type]
 
     cat_stats = df.group_by(cat_col).agg(
         pl.col(target_col).mean().alias("__cat_mean"),
